script now talks...... had to

+++ secureLinux.sh
@@ -23,34 +23,41 @@ update(){
     echo "${PUR}*** Updating system ***${NC}"
 
     ## Full system update
-    echo "Checking for apt_fast (faster)"
-    if [ $(dpkg-query -W -f='${Status}' aptfast 2>/dev/null | grep -c "ok installed") -eq 0 ];
+    espeak "Would you like to completely update now"
+    read -p "Would you like to completely update now (y/n)?" CONT
+    if [ "$CONT" = "y" ]; 
     then
-        echo "apt-fast installed" ## apt-fast
+        echo "Checking for apt_fast (faster)"
+        if [ $(dpkg-query -W -f='${Status}' aptfast 2>/dev/null | grep -c "ok installed") -eq 0 ];
+        then
+            echo "apt-fast installed" ## apt-fast
 
-        apt-fast update
-        apt-fast upgrade
-        apt-fast dist-upgrade
-        apt-fast autoremove
-        apt-fast autoclean
-        apt-fast check
-    else
-        echo "apt-fast not installed" ## apt-get
-        echo "Using default process"
+            apt-fast update
+            apt-fast upgrade
+            apt-fast dist-upgrade
+            apt-fast autoremove
+            apt-fast autoclean
+            apt-fast check
+        else
+            echo "apt-fast not installed" ## apt-get
+            echo "Using default process"
 
-        apt-get update
-        apt-get upgrade
-        apt-get dist-upgrade
-        apt-get autoremove
-        apt-get autoclean
-        apt-get check
+            apt-get update
+            apt-get upgrade
+            apt-get dist-upgrade
+            apt-get autoremove
+            apt-get autoclean
+            apt-get check
+        fi
     fi
 
     ## Automatic updates
+    espeak "Would you like to enable automatic updates"
     read -p "Would you like to enable automatic updates (y/n)?" CONT
     if [ "$CONT" = "y" ]; 
     then
-        echo "Enforcing automatic updates"
+        espeak "Enabling automatic updates"
+        echo "Enabling automatic updates"
         sed -i 's/Update-Package-Lists "0"/Update-Package-Lists "1"/g' /etc/apt/apt.conf.d/20auto-upgrades
     fi
 }
@@ -68,48 +75,61 @@ secure_system(){
     echo "${PUR}*** Securing system ***${NC}"
 
     ## auditd
+    espeak "Would you like to install audit d"
     read -p "Would you like to install auditd (y/n)?" CONT
     if [ "$CONT" = "y" ]; 
     then
+        espeak "Checking for audit d"
         echo "Checking for auditd"
         if [ $(dpkg-query -W -f='${Status}' auditd 2>/dev/null | grep -c "ok installed") -eq 0 ];
         then
+            espeak "Installing auditd"
             echo "Installing auditd"
             apt-get install auditd
             auditd -s enable
         else
+            espeak "audit d already installed"
             echo "auditd already installed"
             auditd -s enable ## enforce it to be enabled
         fi
     fi
 
     ## Clamav
+    espeak "Would you like to install clam a v"
     read -p "Would you like to install clamav (y/n)?" CONT
     if [ "$CONT" = "y" ]; 
     then
+        espeak "Checking for clam a v"
         echo "Checking for clamav"
         if [ $(dpkg-query -W -f='${Status}' clamav 2>/dev/null | grep -c "ok installed") -eq 0 ];
         then
+            espeak "Installing anti virus"
             echo "Installing anti-virus"
             apt-get install clamav
 
+            espeak "Would you like to run a scan now"
             read -p "Would you like to run a scan now (y/n)?" CONT
             if [ "$CONT" = "y" ]; 
             then
+                espeak "Running scan"
                 echo "Running scan (Clamav)"
                 clamscan -r --bell -i /
             fi
         else
+            espeak "Clam a v already installed"
             echo "Clamav already installed"
 
+            espeak "Updating database"
             echo "Updating database (Clamav)"
             /etc/init.d/clamav-freshclam stop ## stop auto-updater so we can update manually
             freshclam
             sudo /etc/init.d/clamav-freshclam start ## restarting auto-updater
 
+            espeak "Would you like run a scan now"
             read -p "Would you like to run a scan now (y/n)?" CONT
             if [ "$CONT" = "y" ]; 
             then
+                espeak "Running scan"
                 echo "Running scan (Clamav)"
                 clamscan -r --bell -i /
             fi
@@ -117,31 +137,40 @@ secure_system(){
     fi
 
     ## rkhunter
+    espeak "Would you like to install r k hunter"
     read -p "Would you like to install rkhunter (y/n)?" CONT
     if [ "$CONT" = "y" ]; 
     then
+        espeak "Checking for r k hunter"
         echo "Checking for rkhunter"
         if [ $(dpkg-query -W -f='${Status}' rkhunter 2>/dev/null | grep -c "ok installed") -eq 0 ];
         then
+            espeak "Installing anti root kit"
             echo "Installing anti-rootkit"
             sudo apt-get install rkhunter
 
+            espeak "Would you like to run a scan now"
             read -p "Would you like to run a scan now (y/n)?" CONT
             if [ "$CONT" = "y" ]; 
             then
+                espeak "Running scan"
                 echo "Running scan (rkhunter)"
                 rkhunter --check
             fi
         else
+            espeak "r k hunter already installed"
             echo "rkhunter already installed"
 
-            echo "updating rkhunter"
+            espeak "Updating r k hunter"
+            echo "Updating rkhunter"
             rkhunter --update
             rkhunter --versioncheck
 
+            espeak "Would you like to run a scan now"
             read -p "Would you like to run a scan now (y/n)?" CONT
             if [ "$CONT" = "y" ]; 
             then
+                espeak "Running scan"
                 echo "Running scan (rkhunter)"
                 rkhunter --check
             fi
@@ -149,9 +178,11 @@ secure_system(){
     fi
 
     ## Compilers
+    espeak "Would you like to disable compilers"
     read -p "Would you like to disable compilers (y/n)?" CONT
     if [ "$CONT" = "y" ]; 
     then
+        espeak "Disabling compilers"
         echo "Disabling compilers"
 
         chmod 000 /usr/bin/as >/dev/null 2>&1
@@ -166,22 +197,31 @@ secure_system(){
     fi
 
     ## debsums
+    espeak "Would you like to install deb sums"
     read -p "Would you like to install debsums (y/n)?" CONT
     if [ "$CONT" = "y" ]; 
     then
+        espeak "Checking for deb sums"
         echo "Checking for debsums"
         if [ $(dpkg-query -W -f='${Status}' debsums 2>/dev/null | grep -c "ok installed") -eq 0 ];
         then
+            espeak "Installing deb sums"
+            echo "Installing deb sums"
+            apt-get install debsums
+
+            espeak "Would you like to run deb sums now"
             read -p "Would you like to run debsums now (y/n)?" CONT
-            if [ "$CONT" = "y" ];
+            if [ "$CONT" = "y" ]; 
             then
                 debsums
             fi
         else
-            apt-get install debsums
+            espeak "deb sums already installed"
+            echo "debsums already installed"
 
+            espeak "Would you like to run deb sums now "
             read -p "Would you like to run debsums now (y/n)?" CONT
-            if [ "$CONT" = "y" ]; 
+            if [ "$CONT" = "y" ];
             then
                 debsums
             fi
@@ -189,9 +229,11 @@ secure_system(){
     fi
 
     ## usbguard
+    espeak "Would you like to install u s b guard"
     read -p "Would you like to install usbguard (y/n)?" CONT
     if [ "$CONT" = "y" ]; 
     then
+        espeak "Checking for u s b guard"
         echo "Checking for usbguard"
         if [ $(dpkg-query -W -f='${Status}' usbguard 2>/dev/null | grep -c "ok installed") -eq 0 ];
         then
@@ -204,6 +246,7 @@ secure_hardware(){
     echo "${PUR}*** Securing hardware access ***${NC}"
 
     ## Insecure IO - thunderbolt
+    espeak "Would you like to disable thunderbolt access"
     read -p "Would you like to disable thunderbolt access (y/n)?" CONT
     if [ "$CONT" = "y" ]; 
     then
@@ -212,13 +255,15 @@ secure_hardware(){
         then
             if ! grep -q 'blacklist thunderbolt' "$File"; 
             then
-                echo "Disabling thunderbolt connections"
+                espeak "Disabling thunder access"
+                echo "Disabling thunderbolt access"
                 echo "blacklist thunderbolt" >> /etc/modprobe.d/thunderbolt.conf
             fi
         fi
     fi
 
     ## Insecure IO - firewire
+    espeak "Would you like to disable fire wire access"
     read -p "Would you like to disable firewire access (y/n)?" CONT
     if [ "$CONT" = "y" ]; 
     then
@@ -227,7 +272,8 @@ secure_hardware(){
         then
             if ! grep -q 'blacklist firewire-core' "$File"; 
             then
-                echo "Disabling firewire"
+                espeak "Disabling fire wire access"
+                echo "Disabling firewire access"
                 echo "blacklist firewire-core" >> /etc/modprobe.d/firewire.conf
             fi
         fi
@@ -238,28 +284,35 @@ secure_connections(){
     echo "${PUR}*** Securing connections ***${NC}"
 
     ## Firewall
+    espeak "Would you like to enable the firewall"
     read -p "Would you like to enable the firewall (y/n)?" CONT
     if [ "$CONT" = "y" ]; 
     then
+        espeak "Enabling the firewall"
         echo "Enabling firewall"
         ufw enable
     fi
 
     ## Insecure protocols - need if statement
+    espeak "Would you like to remove insecure protocols"
     read -p "Would you like to remove insecure protocols (y/n)?" CONT
     if [ "$CONT" = "y" ]; 
     then
+        espeak "Removing insecure protocols"
         echo "Removing insecure protocols"
         sudo apt-get --purge remove xinetd nis yp-tools tftpd atftpd tftpd-hpa telnetd rsh-server rsh-redone-server
     fi
 
     ## psad - need to 'noemail' with context
+    espeak "Would you like to install p s a d"
     read -p "Would you like to install psad (y/n)?" CONT
     if [ "$CONT" = "y" ]; 
     then
+        espeak "Checking for p s a d"
         echo "Checking for psad"
         if [ $(dpkg-query -W -f='${Status}' psad 2>/dev/null | grep -c "ok installed") -eq 0 ];
         then
+            espeak "Installing p s a d"
             echo "Installing psad"
             apt-get install psad
 
@@ -269,17 +322,22 @@ secure_connections(){
             ## Make iptables rules persistent
             if [ $(dpkg-query -W -f='${Status}' iptables-persistent 2>/dev/null | grep -c "ok installed") -eq 0 ];
             then
+                espeak "Enforcing persistent i p tables rules"
                 echo "Enforcing persistent iptables rules"
 
                 apt-get install iptables-persistent
                 service iptables-persistent start
             fi
 
+            espeak "Configuring p s a d settings"
             echo "Configuring psad settings"
             sed -i s/_CHANGEME_/$(whoami)/g /etc/psad/psad.conf ## Hostname
             sed -i s/ALERTING_METHODS            ALL/$(whoami)/g /etc/psad/psad.conf ## Alerting method
         else
+            espeak "p s a d already installed"
             echo "psad already installed"
+
+            espeak "Updating p s a d"
             echo "Updating psad"
             
             psad --sig-update
@@ -288,9 +346,11 @@ secure_connections(){
     fi
 
     ## Malicious domains
+    espeak "Would you like to block malicious domains"
     read -p "Would you like to block malicious domains (y/n)?" CONT
     if [ "$CONT" = "y" ]; 
     then
+        espeak "Black listing malicious domains"
         echo "Black-listing malicious domains"
         File="hosts.txt"
         if [ -e "$File" ]; ## hosts.txt check if exists
@@ -300,9 +360,11 @@ secure_connections(){
             then
                 cat hosts.txt >> /etc/hosts
             else
+                espeak "Malicious domains already blocked"
                 echo "Malicious domains already blocked"
             fi
         else ## hosts.txt missing then download
+            espeak "Host file missing. Would you like to download it"
             read -p "Host file missing. Would you like to download it (y/n)?" CONT
             if [ "$CONT" = "y" ]; 
             then
@@ -316,12 +378,15 @@ secure_connections(){
                     then
                         cat hosts.txt >> /etc/hosts
                     else
+                        espeak "Malicious domains already blocked"
                         echo "Malicious domains already blocked"
                     fi
                 else
-                    echo "Failed to find host.txt... Resuming"
+                    espeak "Failed to find host file"
+                    echo "Failed to find host file... Resuming"
                 fi
             else
+                espeak "Resuming"
                 echo "Resuming"
             fi
         fi
@@ -332,25 +397,31 @@ secure_ssh(){
     echo "${PUR}*** Securing SSH ***${NC}"
 
     ## Limit SSH connections
+    espeak "Would you like to limit SSH connections"  
     read -p "Would you like to limit SSH connections (y/n)?" CONT
     if [ "$CONT" = "y" ]; 
     then
-        echo "Limiting ssh connections"
+        espeak "Limiting SSH connections"
+        echo "Limiting SSH connections"
         ufw limit ssh
         ufw limit openssh
     fi
 
     ## Root login
+    espeak "Would you like to disable SSH root login"
     read -p "Would you like to disable SSH root login (y/n)?" CONT
     if [ "$CONT" = "y" ]; 
     then
+        espeak "Checking if root login allowed"
         echo "Checking if root login allowed"
         File="/etc/ssh/sshd_config"
         if ! grep -q 'DenyUsers root' "$File"; 
         then
+            espeak "Disabling root login"
             echo "Disabling root login sshd"
             echo "DenyUsers root" >> /etc/ssh/sshd_config
         else
+            espeak "Root login already disabled"
             echo "Root login already disabled"
         fi
 
@@ -359,34 +430,43 @@ secure_ssh(){
 
     ## SSH port - review as default sshd_config might need to uncomment port number
     ## need check for this one as port number could be already changed to something other than 22 or 3333
+    espeak "Would you like to change the ssh port from default"
     read -p "Would you like to change ssh port from default (y/n)?" CONT
     if [ "$CONT" = "y" ]; 
     then
+        espeak "Checking SSH port number"
         echo "Checking SSH port number"
         if grep -q 'Port 22' "$File"; 
         then
+            espeak "Changing SSH port to 3 3 3 3"
             echo "Changing SSH port to 3333"
             sed -i 's/Port 22/Port 3333/g' /etc/ssh/sshd_config 
         else
+            espeak "SSH port already changed from default"
             echo "SSH port already changed from default"
         fi
     fi
 
     ## fail2ban
+    espeak "Would you like to install fail2ban"
     read -p "Would you like to install fail2ban (y/n)?" CONT
     if [ "$CONT" = "y" ]; 
     then
+        espeak "Checking for fail2ban"
         echo "Checking for fail2ban"
         if [ $(dpkg-query -W -f='${Status}' fail2ban 2>/dev/null | grep -c "ok installed") -eq 0 ];
         then
+            espeak "Installing fail2ban"
             echo "Installing fail2ban"
             sudo apt-get install fail2ban
         else
+            espeak "fail2ban already installed"
             echo "fail2ban already installed"
         fi
     fi
 
     ## X11Forwarding
+    espeak "Would you like to disable X 11 Forwarding"
     read -p "Would you like to disable X11Forwarding (y/n)?" CONT
     if [ "$CONT" = "y" ]; 
     then
@@ -394,6 +474,7 @@ secure_ssh(){
     fi
 
     ## TCPKeepAlive
+    espeak "Would you like to disable TCPKeepAlive"
     read -p "Would you like to disable TCPKeepAlive (y/n)?" CONT
     if [ "$CONT" = "y" ]; 
     then
@@ -401,6 +482,7 @@ secure_ssh(){
     fi
 
     ## AllowTcpForwarding
+    espeak "Would you like to disable AllowTcpForwarding"
     read -p "Would you like to disable AllowTcpForwarding (y/n)?" CONT
     if [ "$CONT" = "y" ]; 
     then
@@ -408,6 +490,7 @@ secure_ssh(){
     fi
 
     ## AllowAgentForwarding
+    espeak "Would you like to disable AllowAgentForwarding"
     read -p "Would you like to disable AllowAgentForwarding (y/n)?" CONT
     if [ "$CONT" = "y" ]; 
     then
@@ -419,20 +502,27 @@ secure_user(){
     echo "${PUR}*** Securing user ***${NC}"
 
     ## Maximum password age - no need for if statement
+    espeak "Would you like to limit password age to 100 days for root"
     read -p "Would you like to limit password age to 100 days for root (y/n)?" CONT
     if [ "$CONT" = "y" ]; 
     then
+        espeak "Setting maximum password age of root"
         echo "Setting maximum password age of root (100 days)"
         chage -M 100 root
     fi
 }
 
+check_espeak(){
+    if [ $(dpkg-query -W -f='${Status}' espeak 2>/dev/null | grep -c "ok installed") -eq 0 ];
+    then
+        echo "${PUR}This is a one time thing... ${NC}"
+        apt-get install espeak
+    fi
+}
+
 sudo_check
-read -p "Would you like to completely update now (y/n)?" CONT
-if [ "$CONT" = "y" ]; 
-then
-    update
-fi
+check_espeak
+update
 secure_system
 secure_connections
 secure_ssh
